수면시간 저장

+++ src/main/java/com/coawesome/domain/UserVO.java
@@ -21,8 +21,8 @@ public class UserVO {
     private int u_b_month;
     private int u_b_date;
     private int u_sex;
-    private int u_w;
-    private int u_h;
+    private float u_w;
+    private float u_h;
     private int p_id;
     private int a_id;
 }
